the simpler the better

+++ wiwa-service-business/wiwa-service-business-impl/src/test/java/sk/janobono/wiwa/business/impl/component/PriceUtilTest.java
@@ -22,7 +22,7 @@ class PriceUtilTest {
 
     @Test
     void countVatValue_EqualsToExpectedResult() {
-        assertThat(priceUtil.countVatValue((BigDecimal) null, null)).isNull();
+        assertThat(priceUtil.countVatValue(null, null)).isNull();
 
         assertThat(priceUtil.countVatValue(BigDecimal.valueOf(100L), BigDecimal.valueOf(20L)))
                 .isEqualTo(BigDecimal.valueOf(120L).setScale(3, RoundingMode.HALF_UP));
@@ -30,7 +30,7 @@ class PriceUtilTest {
 
     @Test
     void countNoVatValue_EqualsToExpectedResult() {
-        assertThat(priceUtil.countNoVatValue((BigDecimal) null, null)).isNull();
+        assertThat(priceUtil.countNoVatValue(null, null)).isNull();
 
         assertThat(priceUtil.countNoVatValue(BigDecimal.valueOf(120L), BigDecimal.valueOf(20L)))
                 .isEqualTo(BigDecimal.valueOf(100L).setScale(3, RoundingMode.HALF_UP));
